custom lists deletion added

+++ app.js
@@ -125,14 +125,32 @@ app.post("/", async function(req, res) {
 
 app.post("/delete", async function(req, res) {
   const checkedItemId = req.body.checkbox;
+  const listName = req.body.listName;
 
-  try {
-    await Item.findByIdAndDelete(checkedItemId);
-    console.log("Successfully deleted checked item.");
-    res.redirect("/");
-  } catch (err) {
-    console.log(err);
+  if (listName === "Today") {
+    try {
+      await Item.findByIdAndDelete(checkedItemId);
+      console.log("Successfully deleted checked item.");
+      res.redirect("/");
+    } catch (err) {
+      console.log(err);
+      res.status(500).send("Error occurred while deleting the item.");
+    }
+  } else {
+    try {
+      await List.findOneAndUpdate(
+        { name: listName },
+        { $pull: { items: { _id: checkedItemId } } }
+      );
+      res.redirect("/" + listName);
+    } catch (err) {
+      console.log(err);
+      res.status(500).send("Error occurred while updating the list.");
+    }
   }
+  
+
+  
 });
